// src/lib/api/models/DailyLogSegment.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Serializer for daily log segments.
 */
export type DailyLogSegment = {
    startTime: string;
    endTime: string;
    startIndex: number;
    endIndex: number;
    rowIndex: number;
    status: string;
    location: string;
    remarks: string;
};